split data and api keys
